ensure inplace update

+++ backend/btrixcloud/operator/crawls.py
@@ -452,7 +452,7 @@ class CrawlOperator(BaseOperator):
         params["cpu"] = pod_info.newCpu or params.get("redis_cpu")
         params["memory"] = pod_info.newMemory or params.get("redis_memory")
         if self.k8s.enable_auto_resize:
-            params["memory_limit"] = float(params["memory"]) * MEM_LIMIT_PADDING
+            params["memory_limit"] = int(float(params["memory"]) * MEM_LIMIT_PADDING)
         else:
             params["memory_limit"] = self.k8s.max_crawler_memory_size
         params["no_pvc"] = crawl.is_single_page
@@ -617,7 +617,7 @@ class CrawlOperator(BaseOperator):
         params["workers"] = workers
         params["save_profile"] = has_profile and (i == 0)
         if self.k8s.enable_auto_resize:
-            params["memory_limit"] = float(params["memory"]) * MEM_LIMIT_PADDING
+            params["memory_limit"] = int(float(params["memory"]) * MEM_LIMIT_PADDING)
         else:
             params["memory_limit"] = self.k8s.max_crawler_memory_size
         params["storage"] = pod_info.newStorage or params.get("crawler_storage")
